requirements.txt
Django~=3.0.6
djangorestframework~=3.11.0
django-timezone-field~=4.0
arrow~=0.15.6
Jinja2~=2.11.2
requests~=2.23.0
pytz~=2020.1
PyYAML~=5.3.1
psycopg2-binary